refactor to independent components files and add basic app to dev
