+++ allen_oephys_to_nwb/dashboard/allen_oephys_dashboard.py
@@ -1,11 +1,7 @@
 from nwbwidgets.utils.timeseries import (get_timeseries_maxt, get_timeseries_mint,
                                          timeseries_time_to_ind, get_timeseries_in_units,
                                          get_timeseries_tt)
-from tifffile import imread, TiffFile
-from pathlib import Path, PureWindowsPath
 import numpy as np
-import json
-from nwbwidgets.ophys import compute_outline
 
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
@@ -14,232 +10,11 @@ import dash_html_components as html
 import dash_bootstrap_components as dbc
 from dash.dependencies import Input, Output, State, ALL
 import dash
-
-from textwrap import dedent as d
+from pathlib import Path
 import pynwb
 from nwb_web_gui.dashapps.utils.make_components import FileBrowserComponent
-
-
-class TimeControllerComponent(html.Div):
-    """Controller of start time and duration for time series windows"""
-    def __init__(self, parent_app, start=True, duration=True, frame=False,
-                 tmin=0, tmax=1, tstart=0, tduration=1):
-        super().__init__([])
-        self.parent_app = parent_app
-        self.tmax = tmax
-        self.tmax_duration = tmax
-
-        # Start controller'
-        if start:
-            self.slider_start = dcc.Slider(
-                id="slider_start_time",
-                min=tmin, max=tmax, value=tstart, step=0.05,
-            )
-
-            group_start = dbc.FormGroup(
-                [
-                    dbc.Row(
-                        dbc.Col(
-                            dbc.Label('start (s): ' + str(tstart), id='slider_start_label'),
-                            width={'size':12},
-                        ),
-                    ),
-                    dbc.Row(
-                        dbc.Col(
-                            self.slider_start, 
-                            width={'size': 12},
-                        )
-                    )
-                ],
-            )
-
-            @self.parent_app.callback(
-                Output(component_id='slider_start_label', component_property='children'),
-                [Input(component_id='slider_start_time', component_property='value')]
-            )
-            def update_slider_start_label(select_start_time):
-                """Updates Start slider controller label"""
-                return 'start (s): ' + str(select_start_time)
-
-        # Duration controller
-        if duration:
-            self.input_duration = dcc.Input(
-                id="input_duration",
-                type='number',
-                min=.5, max=100, step=.1, value=tduration
-            )
-
-            group_duration = dbc.FormGroup(
-                [
-                    dbc.Row(
-                        dbc.Col(
-                            dbc.Label('duration (s):'), 
-                            width={'size': 12}
-                        )
-                    ),
-                    dbc.Row(
-                        dbc.Col(
-                            self.input_duration,
-                            width={'size':12},
-                        )
-                    )
-                ],
-            )
-
-        # Controllers main layout
-        self.children = dbc.Col([
-            dbc.FormGroup(
-                [
-                    dbc.Col(group_start, width=9),
-                    dbc.Col(group_duration, width=3)
-                ],
-                row=True
-            ),
-            html.Div(id='external-update-max-time-trigger', style={'display': 'none'})
-        ])
-
-        @self.parent_app.callback(
-            [Output('input_duration', 'max'), Output('slider_start_time', 'max')],
-            [Input('external-update-max-time-trigger', 'children'), Input('input_duration', 'value')],
-            [State('input_duration', 'value')]
-        )
-        def update_max_times(trigger, trigger_duration, duration):
-            """
-            Update max slider value when duration change
-            Update max duration and max slider value when new nwb tmax are defined
-            """
-            duration_tmax = self.tmax
-            slider_tmax = self.tmax - duration
-
-            return duration_tmax, slider_tmax
-
-
-class TiffImageSeriesDiv(html.Div):
-    """Div containing tiff image graph and pixelmask button"""
-    def __init__(self, id, parent_app, imageseries, path_external_file=None, pixel_mask=None,
-                 foreign_time_window_controller=None):
-        super().__init__()
-
-        self.graph = TiffImageSeriesGraphComponent(id=id, parent_app=parent_app, imageseries=imageseries, path_external_file=path_external_file, pixel_mask=pixel_mask,foreign_time_window_controller=foreign_time_window_controller)
-        self.pixelmask_btn = dbc.Button('Pixel Mask', id={'type': 'pixelmask_button', 'index': f'mask_btn_{id}'})
-
-        self.children = dbc.Row([
-            dbc.Col(
-                dbc.Card([
-                    dbc.CardBody([
-                        self.graph,
-                        self.pixelmask_btn
-                    ], style={'justify-content': 'center', 'text-align': 'center'}),
-                ]),
-                width={'size': 12},
-            ),
-        ])
-
-
-class TiffImageSeriesGraphComponent(dcc.Graph):
-    """Component that renders specific frame of a Tiff file"""
-    def __init__(self, parent_app, imageseries, path_external_file=None, pixel_mask=None,
-                 foreign_time_window_controller=None, id='tiff_image_series'):
-        super().__init__(id=id, figure={}, config={'displayModeBar': False})
-        self.parent_app = parent_app
-        self.imageseries = imageseries
-        self.pixel_mask = pixel_mask
-
-        if foreign_time_window_controller is not None:
-            self.time_window_controller = foreign_time_window_controller
-        else:
-            self.time_window_controller = None
-
-        # Set controller
-        if foreign_time_window_controller is None:
-            tmin = get_timeseries_mint(imageseries)
-            tmax = get_timeseries_maxt(imageseries)
-            # self.time_window_controller = StartAndDurationController(tmax, tmin)
-        else:
-            self.time_window_controller = foreign_time_window_controller
-
-        # Make figure component
-        if path_external_file is not None:
-            self.tiff = TiffFile(path_external_file)
-            self.n_samples = len(self.tiff.pages)
-            self.page = self.tiff.pages[0]
-            self.n_y, self.n_x = page.shape
-
-            # Read first frame
-            self.image = imread(path_external_file, key=0)
-        else:
-            self.image = []
-            self.tiff = None
-
-        self.out_fig = go.Figure(
-            data=go.Heatmap(
-                z=self.image,
-                colorscale='gray',
-                showscale=False,
-            ), 
-        )
-        self.out_fig.update_layout(
-            xaxis=go.layout.XAxis(showticklabels=False, ticks=""),
-            yaxis=go.layout.YAxis(showticklabels=False, ticks=""),
-        )
-
-        self.figure = self.out_fig
-
-
-    def update_image(self, pos, nwb, relative_path):
-        """Update tiff image frame"""
-
-        frame_number = int(pos * nwb.acquisition['raw_ophys'].rate)
-        path_external = str(Path(relative_path).parent / Path(nwb.acquisition['raw_ophys'].external_file[0]))
-        path_external_file = get_fix_path(path_external)
-
-        if self.tiff is None:
-            self.tiff = TiffFile(path_external_file)
-            self.n_samples = len(self.tiff.pages)
-            self.page = self.tiff.pages[0]
-            n_y, n_x = self.page.shape
-            self.pixel_mask = nwb.processing['ophys'].data_interfaces['image_segmentation'].plane_segmentations['plane_segmentation'].pixel_mask[:]
-
-            mask_matrix = np.zeros((n_y, n_x))
-            for px in self.pixel_mask:
-                mask_matrix[px[1], px[0]] = 1
-
-            self.mask_x_coords, self.mask_y_coords = compute_outline(image_mask=mask_matrix, threshold=0.9)
-
-        self.image = imread(path_external_file, key=frame_number)
-        self.out_fig.data[0].z = self.image
-
-        self.out_fig.update_layout(
-            autosize=False,
-            margin=dict(
-                l=0,
-                r=0,
-                b=10,
-                t=0,
-                pad=0
-            ),
-            height=380
-        )
-
-    def update_pixelmask(self):
-        """ Update pixel mask on self figure """
-
-        if len(self.out_fig.data) == 1:
-            trace = go.Scatter(
-                x=self.mask_x_coords,
-                y=self.mask_y_coords,
-                fill='toself',
-                mode='lines',
-                line={"color": "rgb(219, 59, 59)", "width": 2},
-            )
-            self.out_fig.add_trace(trace)
-        else:
-            if self.out_fig.data[1].x == self.mask_x_coords and self.out_fig.data[1].y == self.mask_y_coords:
-                self.out_fig.data[1].x = []
-                self.out_fig.data[1].y = []
-            else:
-                self.out_fig.data[1].x = self.mask_x_coords
-                self.out_fig.data[1].y = self.mask_y_coords
+from components.time_controller import TimeControllerComponent
+from components.tiff_component import TiffImageSeriesComponent
 
 
 class AllenDashboard(html.Div):
@@ -259,7 +34,6 @@ class AllenDashboard(html.Div):
 
         self.filebrowser = FileBrowserComponent(parent_app=parent_app, id_suffix='allen-dash')
 
-    # def start_dashboard(self):
         if self.path_nwb is not None:
             self.render_dashboard()
 
@@ -318,6 +92,12 @@ class AllenDashboard(html.Div):
             ]
         )
         def update_traces(select_start_time, select_duration):
+
+            ctx = dash.callback_context
+            trigger_source = ctx.triggered[0]['prop_id'].split('.')[1]
+
+            if not trigger_source:
+                raise dash.exceptions.PreventUpdate
 
             time_window = [select_start_time, select_start_time + select_duration]
 
@@ -387,14 +167,16 @@ class AllenDashboard(html.Div):
         )
         def load_nwb_file(click, path):
             if click:
-                if Path(path).is_file() and path.endswith('.nwb'):
-                    self.path_nwb = path
+                if Path(Path(self.parent_app.server.config['DATA_PATH']).parent / path).is_file() and path.endswith('.nwb'):
+                    self.path_nwb = str(Path(self.parent_app.server.config['DATA_PATH']).parent / path)
                     self.render_dashboard()
 
                     display = {'display': 'block'}
                     self.controller_time.tmax = self.controller_tmax
 
                     return 1, self.controller_tmin, display, self.photon_series, str(np.random.rand())
+            else:
+                raise dash.exceptions.PreventUpdate()
 
         @self.parent_app.callback(
             [Output(component_id='figure_photon_series', component_property='figure')],
@@ -496,11 +278,20 @@ class AllenDashboard(html.Div):
             'showline': True,
             'linecolor': 'rgb(0, 0, 0)'
         })
-        self.traces.update_yaxes(title_text="Ephys [V]", row=1, col=1)
-        self.traces.update_yaxes(title_text="dF/F", row=3, col=1)
-        
+        self.traces.update_yaxes(
+            patch={
+                "title": {"text": "Ephys [V]", "font": {"color": "#151733", "size": 16}}
+            },
+            row=1, col=1
+            )
+        self.traces.update_yaxes(
+            patch={
+                "title": {"text": "dF/F", "font": {"color": "#151733", "size": 16}}
+            },
+            row=3, col=1)
+
         self.traces.update_yaxes(patch={
-            "title": {"text": "Spikes", "font": {"color": "black"}},
+            "title": {"text": "Spikes", "font": {"color": "#151733", 'size': 16}},
             "showticklabels": True,
             "ticks": "outside",
             "tickcolor": "white",
@@ -510,11 +301,8 @@ class AllenDashboard(html.Div):
             row=2, col=1
         )
 
-        path_external = str(Path(self.path_nwb).parent / Path(self.nwb.acquisition['raw_ophys'].external_file[0]))
-        path_external = get_fix_path(path_external)
-
         # Two photon imaging
-        self.photon_series = TiffImageSeriesDiv(
+        self.photon_series = TiffImageSeriesComponent(
             id='figure_photon_series',
             parent_app=self.parent_app,
             imageseries=self.nwb.acquisition['raw_ophys'],
@@ -528,7 +316,6 @@ class AllenDashboard(html.Div):
             margin=dict(l=10, r=10, t=70, b=70),
             # width=300, height=300,
         )
-
 
     def update_spike_traces(self, time_window):
         """Updates list of go.Scatter objects at spike times"""
@@ -547,11 +334,3 @@ class AllenDashboard(html.Div):
                 mode='lines'),
                 row=2, col=1
             )
-
-
-def get_fix_path(path):
-    if '\\' in path:
-        win_path = PureWindowsPath(path)
-        return Path(win_path)
-    else:
-        return Path(path)
